Fix Test's Automation
