+++ Spinoza_Automation/.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/Spinoza/2022/5/23/refactorings.history
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<session version="1.0">&#x0A;<refactoring comment="Rename type &apos;Test3&apos; to &apos;Test5&apos;&#x0D;&#x0A;- Original project: &apos;Spinoza&apos;&#x0D;&#x0A;- Original element: &apos;Test3&apos;&#x0D;&#x0A;- Renamed element: &apos;Test5&apos;&#x0D;&#x0A;- Update references to refactored element&#x0D;&#x0A;- Update textual occurrences in comments and strings" description="Rename type &apos;Test3&apos;" flags="589830" id="org.eclipse.jdt.ui.rename.type" input="/src=/optional=/true=/=/maven.pomderived=/true=/&lt;{Test3.java[Test3" matchStrategy="1" name="Test5" qualified="false" references="true" similarDeclarations="false" stamp="1653816176447" textual="false" version="1.0"/>&#x0A;<refactoring comment="Rename type &apos;Test4&apos; to &apos;Test3&apos;&#x0D;&#x0A;- Original project: &apos;Spinoza&apos;&#x0D;&#x0A;- Original element: &apos;Test4&apos;&#x0D;&#x0A;- Renamed element: &apos;Test3&apos;&#x0D;&#x0A;- Update references to refactored element&#x0D;&#x0A;- Update textual occurrences in comments and strings" description="Rename type &apos;Test4&apos;" flags="589830" id="org.eclipse.jdt.ui.rename.type" input="/src=/optional=/true=/=/maven.pomderived=/true=/&lt;{Test4.java[Test4" matchStrategy="1" name="Test3" qualified="false" references="true" similarDeclarations="false" stamp="1653816190307" textual="false" version="1.0"/>&#x0A;<refactoring comment="Rename type &apos;Test5&apos; to &apos;Test4&apos;&#x0D;&#x0A;- Original project: &apos;Spinoza&apos;&#x0D;&#x0A;- Original element: &apos;Test5&apos;&#x0D;&#x0A;- Renamed element: &apos;Test4&apos;&#x0D;&#x0A;- Update references to refactored element&#x0D;&#x0A;- Update textual occurrences in comments and strings" description="Rename type &apos;Test5&apos;" flags="589830" id="org.eclipse.jdt.ui.rename.type" input="/src=/optional=/true=/=/maven.pomderived=/true=/&lt;{Test5.java[Test5" matchStrategy="1" name="Test4" qualified="false" references="true" similarDeclarations="false" stamp="1653816260405" textual="false" version="1.0"/>
+</session>
