+++ Spinoza_Automation/.metadata/.plugins/org.eclipse.ltk.core.refactoring/.refactorings/Spinoza/2022/5/23/refactorings.index
@@ -1,0 +1,3 @@
+1653816176447	Rename type 'Test3'
+1653816190307	Rename type 'Test4'
+1653816260405	Rename type 'Test5'
