+++ Spinoza_Automation/.metadata/version.ini
@@ -1,3 +1,3 @@
-#Sat May 28 18:41:27 IDT 2022
+#Sun May 29 12:00:54 IDT 2022
 org.eclipse.core.runtime=2
 org.eclipse.platform=4.22.0.v20211124-1800
