+++ Spinoza_Automation/Spinoza/src/Startup.java
@@ -1,5 +1,4 @@
 import java.util.concurrent.TimeUnit;
-
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 
@@ -10,16 +9,15 @@ public class Startup {
 		WebDriver driver = new ChromeDriver();
 		driver.manage().timeouts().implicitlyWait(15, TimeUnit.SECONDS);
 
-		
 		driver.manage().window().maximize();
 		driver.get("http://localhost:3000/");
 //		driver.get("https://wonderful-bay-0dbd0a003.1.azurestaticapps.net/");
-		
-		//TEST'S CREATE TEST
+
+		// TEST'S CREATE TEST
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		System.out.println("TEST START - TEST'S CREATE");
+		System.out.println("TEST START - TEST'S CREATE TEST");
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		
+
 		Test1 t1 = new Test1();
 		t1.CreateNewTest(driver); // CREATE TEST AND CHECK TITLE CONFLICT
 		Thread.sleep(1000);
@@ -29,18 +27,18 @@ public class Startup {
 		Thread.sleep(1000);
 		t1.CreateNewTest_Test3(driver); // CREATE TEST WITH ALL DATA FIELDS
 		Thread.sleep(1000);
-		
+
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		System.out.println("TEST END - TEST'S CREATE");
+		System.out.println("TEST END - TEST'S CREATE TEST");
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
 
 		Thread.sleep(5000);
-		
-		//TEST'S EDIT TEST
+
+		// TEST'S EDIT TEST
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		System.out.println("TEST START - TEST'S EDIT");
+		System.out.println("TEST START - TEST'S EDIT TEST");
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		
+
 		Test2 t2 = new Test2();
 		t2.EditTestTitle(driver); // EDIT TEST TITLE
 		Thread.sleep(1000);
@@ -48,47 +46,64 @@ public class Startup {
 		Thread.sleep(1000);
 		t2.EditTestDescription(driver); // EDIT TEST DESCRIPTION
 		Thread.sleep(1000);
-		t2.EditTestTags(driver); // EDIT TAGS
+		t2.EditTestTags(driver); // EDIT TEST TAGS
 		Thread.sleep(1000);
 		t2.EditTestQuestions_CreateNewQuestion(driver); // EDIT TEST BY CREATE NEW QUESTION
 		Thread.sleep(1000);
 		t2.EditTestQuestion_ChooseFromCatalog(driver); // EDIT TEST BY CHOSING QUESTION FORM CATALOG
 		Thread.sleep(1000);
-		
+
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		System.out.println("TEST END - TEST'S EDIT");
+		System.out.println("TEST END - TEST'S EDIT TEST");
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
 
 		Thread.sleep(5000);
-		
-		//TEST'S CREATE QUESTION
-		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		System.out.println("TEST START - ADD QUESTION TO TEST");
-		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		
-		Test3 t3 = new Test3();
-		t3.CreateNewQuestion(driver, 1); // CREATE QUESTION WITH 'TEXT' ANSWERS
-		Thread.sleep(1000);
-		t3.CreateNewQuestion(driver, 2); // CREATE QUESTION WITH 'MULTIPLE' ANSWERS
-		Thread.sleep(1000);
-		
-		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		System.out.println("TEST END - ADD QUESTION TO TEST");
-		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
 
-		Thread.sleep(5000);
-		
-		//TEST'S SEARCH TEST BY TAG
+		// TEST'S SEARCH TEST BY TAG
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
 		System.out.println("TEST START - SEARCH TEST'S BY TAG");
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
-		
-		Test4 t4 = new Test4();
-		t4.SearchTestsByTag(driver, "Python");
+
+		Test3 t3 = new Test3();
+		t3.SearchTestsByTag(driver, "Python");
 		Thread.sleep(1000);
-		
+
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
 		System.out.println("TEST END - SEARCH TEST'S BY TAG");
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+
+		// TEST'S CREATE QUESTION
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+		System.out.println("TEST START - TEST'S CREATE QUESTION");
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+
+		Test4 t4 = new Test4();
+		t4.CreateNewQuestion(driver, 1); // CREATE QUESTION WITH 'TEXT' ANSWERS
+		Thread.sleep(1000);
+		t4.CreateNewQuestion(driver, 2); // CREATE QUESTION WITH 'MULTIPLE' ANSWERS
+		Thread.sleep(1000);
+
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+		System.out.println("TEST END - TEST'S CREATE QUESTION");
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+
+		Thread.sleep(5000);
+
+		// TEST'S EDIT QUESTION
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+		System.out.println("TEST START - TEST'S EDIT QUESTION");
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+
+		Test5 t5 = new Test5();
+		t5.EditQuestionTitle(driver); // EDIT QUESTION TITLE
+		Thread.sleep(1000);
+		t5.EditQuestionDescription(driver); // EDIT QUESTION DESCRIPTION
+		Thread.sleep(1000);
+		t5.EditQuestionTags(driver); // EDIT QUESTION TAGS
+		Thread.sleep(1000);
+
+		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
+		System.out.println("TEST END - TEST'S EDIT QUESTION");
 		System.out.println("-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-");
 
 		driver.quit();
