+++ Spinoza_Automation/Spinoza/src/Test2.java
@@ -1,7 +1,6 @@
 import java.util.Date;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
-
 import org.openqa.selenium.Alert;
 import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
